+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #564 - new algorithm for aligning and stacking
 - #556 - store mag limit in catalog files
 - #553 - store std coefficients in catalog files
+- #547 - make sequence and combined photometry together 
 - #545 - preserve records even mag is below limit
 - #539 - save graphs with --saveGraph option
 - #537 - update reference catalog
@@ -27,6 +28,3 @@
 - #449 - manage hot/dead pixels
 
 **1.0 - Original release** 
-
-
-